Work with mixed configuration
The incoming data can be formatted ni different ways depending on
the weather softare. Some send a query string with a name mapping to
the data value, some send a space separated string.

for space separated strings, we map the fields in the string by
field number.

To handle cases where the configuration may contain mixed field number
and query string name, first verify that the value is an integer
before trying to use it as a field index.