Add an empty nodedef file.
Polyglot seems to want to install the profile files before it starts
the node server. However, we want to have the node server create the
node definitions based on user configuration. A bit of a catch 22. So
create an empty nodedef file and allow that to be initially installed.
It should be overwritten with a valid nodedef when the node server
starts.