Set initial controller status to true.
There seems to be a bug in the way the controller online status
is handled with reportDriver. If the controller defines the initial
value as 0 (offline), it will be switched to 1 (online) when the
controller is connected to Polyglot. However, when queried and
reportDriver is called, it returns the initial value of 0 (offline)
and it then stays offline until the node server is restared.

To work around this, we set the initial value to 1. Then when
queried, it returns the correct value.